% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS_from_xyz.R
\name{LAS_from_xyz}
\alias{LAS_from_xyz}
\title{Transform raw coordinates into an object from the LAS class}
\usage{
LAS_from_xyz(xyz)
}
\arguments{
\item{xyz}{data.frame or matrix with 3D coordinates (xyz).}
}
\value{
An object from the \code{\linkS4class{LAS}}.
}
\description{
Transform raw coordinates into an object from the LAS class
}
\examples{
require(lidR)

x <- runif(1000, 0, 50)
y <- runif(1000, 0, 50)
z <- runif(1000, 0, 20)

las <- LAS_from_xyz(data.frame(x, y, z))
las
plot(las)




}
\seealso{
\code{\link[lidR]{writeLAS}}
}
