% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2012.R
\name{var2012}
\alias{var2012}
\title{Varhola et al. (2012)'s method}
\usage{
var2012(
  X,
  Y,
  Z,
  station_point,
  lens_height = 1,
  sphere_diameter = 0.15,
  minimum_base_constant = 0.215,
  distance_thr = 0.75,
  image_size = 750,
  filename
)
}
\arguments{
\item{X}{numeric vector. X coordinates of 3D points (a point cloud).}

\item{Y}{numeric vector. Y coordinates of 3D points (a point cloud).}

\item{Z}{numeric vector. Z coordinates of 3D points (a point cloud).}

\item{lens_height}{numeric. Single value. Height above ground of the virtual lens.}

\item{sphere_diameter}{numeric. Single value. Diameter in meters.}

\item{minimum_base_constant}{numeric. Single value. It controls the minimum
size to be printed. See reference for details.}

\item{distance_thr}{numeric. Single value. Minimum allowed distance to virtual lens.}

\item{image_size}{numeric. Single value. Width in pixels}

\item{filename}{character. Argument to pass to \code{\link[grDevices]{jpeg}}.}

\item{station_poin}{numeric vector. X, Y, and Z coordinates of a 3D point. It
represents the position of the virtual tripod.}
}
\value{
None. It writes a JPEG on \code{filename}.
}
\description{
Varhola et al. (2012)'s method
}
\examples{
require(lidR)
path <- system.file("external/745_r75m_d1m_local.las", package = "hpsynt")
las <- readLAS(path)

filename <- tempfile("var2012", fileext = ".jpg")
var2012(las@data$X, las@data$Y, las@data$Z,
       lens_height = 1,
       sphere_diameter = 0.5,
       minimum_base_constant = 0.0215,
       distance_thr = 0.75,
       image_size = 2000,
       filename = filename
       )
plotRGB(brick(filename))
unlink(filename)
}
\references{
Varhola A, Frazer GW, Teti P, Coops NC (2012). Estimation of forest
structure metrics relevant to hydrologic modeling using coordinate
transformation of airborne laser scanning data. Hydrology and Earth System Sciences
Discussions. 9: 5531-5573. - doi: 10.5194/hessd-9-5531-2012
}
\seealso{
\code{\link{moe2014}}
}
