% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe2014.R
\name{moe2014}
\alias{moe2014}
\title{Moese et al. (2014)'s method}
\usage{
moe2014(
  X,
  Y,
  Z,
  lens_height = 1,
  size_weight = 0.5,
  image_size = 1000,
  filename
)
}
\arguments{
\item{X}{numeric vector. X coordinates of 3D points (a point cloud).}

\item{Y}{numeric vector. Y coordinates of 3D points (a point cloud).}

\item{Z}{numeric vector. Z coordinates of 3D points (a point cloud).}

\item{lens_height}{numeric. Single value. Height above ground of the virtual lens.}

\item{size_weight}{numeric. Single value. It controls the printed sizes. This parameter
was introduced since the original method uses Matlab to print (see reference), which
seems to manage point size different than R.}

\item{image_size}{numeric. Single value. Width in pixels}

\item{filename}{character. Argument to pass to \code{\link[grDevices]{jpeg}}.}
}
\value{
None. It writes a JPEG on \code{filename}.
}
\description{
Moese et al. (2014)'s method
}
\references{
Moeser D, Roubinek J, Schleppi P, Morsdorf F, Jonas T (2014). Canopy closure,
LAI and radiation transfer from airborne LiDAR synthetic images. Agricultural and
Forest Meteorology. 197: 158-168. - doi: 10.1016/j.agrformet.2014.06.008
}
\seealso{
\code{\link{var2012}}
}
